Revert "microbit2 제거, microbit으로 통합"
This reverts commit 175d9214ebe7f9aa2d6e64d73f94ccdb000bbe69.
