+++ app/modules/microbit2.json
@@ -1,0 +1,36 @@
+{
+    "id": "FF0101",
+    "name": {
+        "en": "Microbit2",
+        "ko": "마이크로빗빗"
+    },
+    "category": "board",
+    "platform": ["win32", "darwin"],
+    "icon": "microbit2.png",
+    "module": "microbit2.js",
+    "url": "https://microbit.org/ko/",
+    "select_com_port": true,
+    "reconnect": true,
+    "tryFlasherNumber": 1,
+    "firmware": {
+        "type": "copy",
+        "name": "microbit",
+        "afterDelay": 3000
+    },
+    "entry": {
+        "protocol": "json"
+    },
+    "driver": {
+        "win32-ia32": "mbedWinSerial/mbedWinSerial_16466.exe",
+        "win32-x64": "mbedWinSerial/mbedWinSerial_16466.exe"
+    },
+    "hardware": {
+        "type": "serial",
+        "control": "slave",
+        "byteDelimiter": [13, 10],
+        "duration": 64,
+        "vendor": "",
+        "baudRate": 115200,
+        "bufferSize": 16
+    }
+}
